requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
